backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
playwright==1.49.1
python-dotenv==1.0.1
schedule==1.2.2

